requirements.txt
Flask==3.0.2
gunicorn==23.0.0
requests==2.31.0
pyfiglet==0.8.post1
Pillow==10.0.0
python-dotenv==1.0.1
